+++ pkg/providers/cloudaccess/cloudaccess.go
@@ -1,0 +1,7 @@
+package cloudaccess
+
+import "github.com/akamai/terraform-provider-akamai/v6/pkg/providers/registry"
+
+func init() {
+	registry.RegisterSubprovider(NewSubprovider())
+}