+++ pkg/providers/cloudaccess/cloudaccess_errors.go
@@ -1,0 +1,8 @@
+package cloudaccess
+
+import "errors"
+
+var (
+	// ErrCloudAccessKey is returned when operation on cloud access key fails
+	ErrCloudAccessKey = errors.New("cloud access key")
+)